+++ src/youwol/pipelines/pipeline_python_backend/template.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from typing import NamedTuple
 
 # Youwol
-from youwol import __release_version__
+from youwol import __version__
 
 # Youwol application
 from youwol.app.environment import ProjectTemplate
@@ -106,7 +106,7 @@ async def generate_template(folder: Path, parameters: dict[str, str], context: C
             for source, replacement in [
                 ["package_name", inputs.name],
                 ["default_port", str(inputs.port)],
-                ["minimal_youwol_version", __release_version__],
+                ["youwol_version", __version__],
             ]:
                 sed_inplace(file_to_sed, "{{" + source + "}}", replacement)
 
@@ -134,7 +134,7 @@ async def generate_template(folder: Path, parameters: dict[str, str], context: C
             dst = package_folder / file.name.replace(".txt", "")
             shutil.copyfile(
                 src=file,
-                dst=package_folder / file.name.replace(".txt", ""),
+                dst=dst,
             )
             replace_patterns(dst)
 
@@ -147,11 +147,13 @@ async def generate_template(folder: Path, parameters: dict[str, str], context: C
 
         # Backend API scripts
 
-        for file in ["install.sh", "start.sh"]:
+        for file in ["install.sh", "start.sh", "Dockerfile.txt"]:
+            dst = project_folder / file.replace(".txt", "")
             shutil.copyfile(
                 src=Path(__file__).parent / "template" / file,
-                dst=project_folder / file,
+                dst=dst,
             )
+            replace_patterns(dst)
 
         # Pipeline definition, need to be at the end (such that `ProjectFindersImpl` can 'auto-discover'
         # a valid project when `watch` is `True`).
@@ -160,7 +162,6 @@ async def generate_template(folder: Path, parameters: dict[str, str], context: C
             src=src_template_folder / ".yw_pipeline" / "yw_pipeline.py",
             dst=pipeline_folder / "yw_pipeline.py",
         )
-
         return parameters["name"], project_folder